Create ProxyConnectionFactory
To align with builder pattern usage in other r2dbc projects:
- Create `ProxyConnectionFactory`
- Move `ProxyConnectionFactoryBuilder` to `ProxyConnectionFactory.Builder`
- Use `builder()` and `create()` methods.
